requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
pydantic==2.8.2
python-multipart==0.0.9
prometheus-client==0.20.0
Jinja2==3.1.4
python-dotenv==1.0.1
python-jose[cryptography]==3.5.0

